refactored CarController to follow SOLID principle
