+++ src/main/resources/templates/carList.html
@@ -10,7 +10,7 @@
 <div class="container my-2">
   <h2>Car' List</h2>
   <a th:href="@{/car/createCar}" class="btn btn-primary btn-sm mb-3">Create Car</a>
-
+  <a th:href="@{/}" class="btn btn-primary btn-sm mb-3">Go Back to Homepage</a>
   <table border="1" class="table table-striped table-responsive-md">
     <thead>
     <tr>
