+++ src/main/resources/templates/home.html
@@ -12,6 +12,7 @@
     <p>The most reliable online shop for all your needs.</p>
 
     <a th:href="@{/product/list}" class="btn btn-primary btn-lg mt-3">Go to Product List</a>
+    <a th:href="@{/car/listCar}" class="btn btn-primary btn-lg mt-3">Go to Car List</a>
 </div>
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
 </body>
